Sol version up and converting to hardhat
