+++ compileHasher.js
@@ -6,16 +6,19 @@ const genContract = require('circomlib/src/poseidon_gencontract.js')
 
 // where Truffle will expect to find the results of the external compiler
 // command
-const outputPath = path.join(__dirname, 'build', 'Hasher.json')
+const outputPath = path.join(__dirname, 'artifacts/contracts/MerkleTreeWithHistory.sol', 'Hasher.json')
 
 function main () {
-  const contract = {
-    contractName: 'Hasher',
-    abi: genContract.generateABI(2),
-    bytecode: genContract.createCode(2)
-  }
+  const data = fs.readFileSync(outputPath)
+  let contract = JSON.parse(data);
+  contract.bytecode = genContract.createCode('mimcsponge', 220);
+  // const contract = {
+  //   contractName: 'Hasher',
+  //   abi: genContract.abi,
+  //   bytecode: genContract.createCode('mimcsponge', 220)
+  // }
 
-  fs.writeFileSync(outputPath, JSON.stringify(contract, null, 2))
+  fs.writeFileSync(outputPath, JSON.stringify(contract))
 }
 
 main()
