+++ contracts/Mocks/BadRecipient.sol
@@ -1,6 +1,5 @@
 // SPDX-License-Identifier: MIT
-pragma solidity ^0.6.0;
-pragma experimental ABIEncoderV2;
+pragma solidity ^0.8.0;
 
 contract BadRecipient {
   fallback() external {
