+++ contracts/Mocks/IUSDT.sol
@@ -1,6 +1,5 @@
 // SPDX-License-Identifier: MIT
-pragma solidity ^0.6.0;
-pragma experimental ABIEncoderV2;
+pragma solidity ^0.8.0;
 
 abstract contract ERC20Basic {
     uint public _totalSupply;
