+++ contracts/Verifier.sol
@@ -120,17 +120,17 @@ contract Verifier {
     }
 
     function verifyingKey() internal pure returns (VerifyingKey memory vk) {
-        vk.alfa1 = Pairing.G1Point(uint256(21319828715501805369083478299296312268707688624392371715189749747934301976886), uint256(7075751588106346862968061466078032096364481862986616788417769258216859729925));
-        vk.beta2 = Pairing.G2Point([uint256(7491930141116389898451232062082175006524927477931829070088802383527997871706), uint256(11645867074280589749191204371114330347805550394886484770840486558008119202093)], [uint256(9822273585043667053470147287458424784983564724691292310735901331083600158991), uint256(4668564564936726250404150668607370216371910282149474392608522091403630870643)]);
-        vk.gamma2 = Pairing.G2Point([uint256(9333220218831783177509618794927471219229906345297818180974764883382706425479), uint256(21251337496446074684512973351167621890795826643789706351019627162549969838017)], [uint256(1579277903494891131382545211244064614258505225529865050538682580737498786113), uint256(9889813466913012791717438339986657160907853054272580044649024599294959699031)]);
-        vk.delta2 = Pairing.G2Point([uint256(10366681292672413036012513150228832795603372237850211411817372694706587521712), uint256(9461592606816701418655911720622828060582523593985715417969842436741689504051)], [uint256(14775986697241998036361336069875532314889416931155074123055047444133147289485), uint256(7710569105028722235140817940879395220311153136498376379324017083126058038584)]);
-        vk.IC[0] = Pairing.G1Point(uint256(7527448446822633881006353707476441802473155195337920176002639371616051876310), uint256(16090178411609243205945249838277936183154752783263782754890295557078303008742));
-        vk.IC[1] = Pairing.G1Point(uint256(1483660204145759066689105364889142852378887056649467825565854245277728990091), uint256(17699940123291102574042810053772720396445030403776292850299296347600166706137));
-        vk.IC[2] = Pairing.G1Point(uint256(5342846082419376373994466268921731472096616366119886605553308996945852485033), uint256(8205900749307517433811652766273696638633443682726253107477121806632252389600));
-        vk.IC[3] = Pairing.G1Point(uint256(20147757608706538505801715302295944474887326029708300364266606685362846331931), uint256(17099615898411516079912033935761326554233729393362668445867495674077124447580));
-        vk.IC[4] = Pairing.G1Point(uint256(3153588507625282016043641550064126711491194246207082637322062663611293234501), uint256(15070826965052566015406543410701714030015584363346516828568290969046574739462));
-        vk.IC[5] = Pairing.G1Point(uint256(6258025095766337178815554018844185794859833085619799760446640254440331079351), uint256(20554520564509628557524241017970163965204843667100841561196959113207738810159));
-        vk.IC[6] = Pairing.G1Point(uint256(8778828242319755663615102735137518889855537107242704996009188068317335577189), uint256(8063146526821170984013194268119119545136211989514966324451629476217833100571));
+        vk.alfa1 = Pairing.G1Point(uint256(2716894789766860705480572884882327526272063118273905481829175777148901691147), uint256(20016447410483166612266960252935361813714256564414892628087341330522790124810));
+        vk.beta2 = Pairing.G2Point([uint256(12567836386270923228486247518702590557040361451608923744881470382218031804886), uint256(1748514736653164701983851759173770138133823205513353468366220329373461698915)], [uint256(13269980721619281266903816012300444941991930185938630084015366067182447113623), uint256(1834809592841059405278305596841315232986118611096606326082156528184832722804)]);
+        vk.gamma2 = Pairing.G2Point([uint256(20847594279349311191146772616701905286381119559674308950926661652008716159438), uint256(3730675332352650420414606611132380815992714607847930130167526444202416110090)], [uint256(13767278491054261791304651741228170640855028408095879212804692055918436015783), uint256(14092293549396327020376820007919666163827372453664343035225237454858790406248)]);
+        vk.delta2 = Pairing.G2Point([uint256(4756987968857785630897921354545664436568762945816099581258503347476896612310), uint256(4194628836529550404099720617978926989806366456658385088037797524230988606176)], [uint256(8784053917622543010028964808936782974335449326009225897108868949849645672158), uint256(11698968986383822975066216413396782493134967347350832042400230438282097288257)]);
+        vk.IC[0] = Pairing.G1Point(uint256(14100369735059767803679748418345854294209882690098460157004418443935822778215), uint256(15099973429503551806081922905954162228179255542246893007813511025336555197045));
+        vk.IC[1] = Pairing.G1Point(uint256(16550095487700438675080637606031781149993007443904218123904693371876731351709), uint256(10847365755001615928631150237631787042357473604968016110708571053610448654887));
+        vk.IC[2] = Pairing.G1Point(uint256(17297298099889221789066515873675595243406300536031099349748349434565465895769), uint256(20804746862711591831925348252431989025553817808352684688886953900768182880650));
+        vk.IC[3] = Pairing.G1Point(uint256(19541437049444264654461479682530323058044882857912571799678847904759757470054), uint256(7488793752010650206016075121359369249582501901345628849820136842362786597707));
+        vk.IC[4] = Pairing.G1Point(uint256(20284356839878887748570591582156585531660103512280935115506928094121622855108), uint256(393523518660990912548820383661226011369184513838017090087371113759969097244));
+        vk.IC[5] = Pairing.G1Point(uint256(5505734420126423876426876590968916939396746019511709785214062095338198645374), uint256(11264408126063607739547308160151837939814991083501878917464778119186507207159));
+        vk.IC[6] = Pairing.G1Point(uint256(15652768789631477045740995877587608228337567249898304650946018710750058039826), uint256(18252134236204641863939970570255514322669252462067784015422352196620876695675));
 
     }
