+++ package.json
@@ -5,16 +5,12 @@
   "main": "index.js",
   "scripts": {
     "build:circuit": "mkdir -p build/circuits && scripts/buildCircuit.sh",
-    "build:contract": "npx truffle compile",
+    "build:contract": "npx hardhat compile",
     "build:browserify": "npx browserify cli.js -o index.js --exclude worker_threads",
     "build": "npm run build:circuit && npm run build:contract && npm run build:browserify",
     "browserify": "npm run build:browserify",
-    "test": "npx truffle test",
-    "migrate": "npm run migrate:kovan",
-    "migrate:dev": "npx truffle migrate --network development --reset",
-    "migrate:kovan": "npx truffle migrate --network kovan --reset --f 1 --to 4",
-    "migrate:rinkeby": "npx truffle migrate --network rinkeby --reset",
-    "migrate:mainnet": "npx truffle migrate --network mainnet",
+    "test": "npx ./compileHasher.js && npx hardhat test test/hardhatTest.js",
+    "deploy": "npx hardhat run scripts/deploy.js",
     "eslint": "npx eslint --ignore-path .gitignore .",
     "flat": "npx truffle-flattener contracts/ETHSacred.sol > ETHSacred_flat.sol && npx truffle-flattener contracts/ERC20Sacred.sol > ERC20Sacred_flat.sol",
     "download": "node downloadKeys.js"
@@ -23,15 +19,16 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@openzeppelin/contracts": "^3.1.0",
-    "@truffle/contract": "^4.0.39",
-    "@truffle/hdwallet-provider": "^1.0.24",
+    "@nomiclabs/hardhat-ethers": "^2.0.1",
+    "@nomiclabs/hardhat-waffle": "^2.0.1",
+    "ethereum-waffle": "^3.2.2",
+    "ethers": "^5.4.4",
+    "hardhat": "^2.6.0",
+    "@openzeppelin/contracts": "^4.5.0",
     "axios": "^0.19.0",
     "bn-chai": "^1.0.1",
     "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
-    "circom": "0.0.35",
-    "circomlib": "git+https://github.com/tornadocash/circomlib.git#3b492f9801573eebcfe1b6c584afe8a3beecf2b4",
     "fixed-merkle-tree": "^0.3.4",
     "commander": "^4.1.1",
     "dotenv": "^8.2.0",
@@ -39,9 +36,11 @@
     "eslint-config-prettier": "^6.11.0",
     "eslint-plugin-prettier": "^3.1.4",
     "eth-json-rpc-filters": "^4.1.1",
-    "snarkjs": "git+https://github.com/tornadocash/snarkjs.git#869181cfaf7526fe8972073d31655493a04326d5",
-    "truffle": "^5.0.44",
-    "truffle-flattener": "^1.4.2",
-    "websnark": "git+https://github.com/tornadocash/websnark.git#86a526718cd6f6f5d31bdb1fe26a9ec8819f633e"
+    "web3": "^1.2.2",
+    "web3-utils": "^1.2.2",
+    "circom": "0.0.35",
+    "circomlib": "git+https://github.com/tornadocash/circomlib.git#3b492f9801573eebcfe1b6c584afe8a3beecf2b4",
+    "websnark": "git+https://github.com/tornadocash/websnark.git#86a526718cd6f6f5d31bdb1fe26a9ec8819f633e",
+    "snarkjs": "git+https://github.com/tornadocash/snarkjs.git#869181cfaf7526fe8972073d31655493a04326d5"
   }
 }
